Fixed marker on Avatar
