+++ packages/theme/styles/common.scss
@@ -113,6 +113,10 @@
 }
 .normal-font .antiPanel-application .app .icon-container.noty { clip-path: url(#notify-normal); }
 .small-font .antiPanel-application .app .icon-container.noty { clip-path: url(#notify-small); }
+.normal-font .avatar-container.withStatus.x-small-hole:not(.circle) { mask-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'%3E%3Cpath d='M8,12.5C8,10,10,8,12.5,8c1.4,0,2.7,0.7,3.5,1.7V0H0v16h9.7C8.7,15.2,8,13.9,8,12.5z' /%3E%3C/svg%3E"); }
+.small-font .avatar-container.withStatus.x-small-hole:not(.circle) { mask-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 14 14'%3E%3Cpath d='M7,10.9C7,8.8,8.8,7,10.9,7c1.2,0,2.3,0.6,3.1,1.5V0H0v14h8.5C7.6,13.3,7,12.2,7,10.9z' /%3E%3C/svg%3E"); }
+.normal-font .avatar-container.withStatus:is(.small-hole, .medium-hole, .auto-hole):not(.circle) { mask-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'%3E%3Cpath d='M10,13.5c0-1.9,1.6-3.5,3.5-3.5c1,0,1.9,0.4,2.5,1.1V0H0v16h11.1C10.4,15.4,10,14.5,10,13.5z' /%3E%3C/svg%3E"); }
+.small-font .avatar-container.withStatus:is(.small-hole, .medium-hole, .auto-hole):not(.circle) { mask-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 14 14'%3E%3Cpath d='M8.8,11.8c0-1.7,1.4-3.1,3.1-3.1c0.9,0,1.6,0.4,2.2,0.9V0H0v14h9.7C9.1,13.4,8.8,12.7,8.8,11.8z' /%3E%3C/svg%3E"); }
 
 .antiPanel-navigator,
 .antiPanel-component {
